Move models and system extensions to separate files.
